// src/app/layout/models/Event.ts
export interface Event {
    _id?: string;
    code?: string;
    userId?: string;
    hostName?: string;
    name?: string;
    description?: string;
    amount?: number;
    eventDate?: string;
    eventTime?: string;
    duration?: string;
    category?: string;
    image?: string;
    rating?: number;
    free?: boolean;
}